Reorganize type-aware layout files to follow Quartz conventions
- Move typeLayouts.ts to quartz/types/ where it belongs
- Update all import paths to reference new location
- Add comprehensive file index to documentation
- Create CLAUDE.md with project instructions emphasizing:
  - Always leverage built-in Quartz functionality first
  - Strict file organization conventions
  - Type-aware system overview and best practices
  - Common patterns and debugging tips

This ensures all source code follows Quartz directory structure and
provides clear guidance for future development.
